1- admin/admin_util.php => creating previews follows symlinks. e.g. when creating previews of album x/y/z where y is a symlink to S, it will actually generates them for x/S/z.
2- src/common_media.php => when viewing an album, symlinks are looked for in the parent dirs too.
Same here, visiting x/y/z will actually visit x/S/z.
The difference is that before it would only resolve x/y to x/S but would not do it for sub-dirs.

In both cases resolving symlinks is done only once (meaning is x/S/z also has a symlink somewhere,
it won't get resolved.) Otherwise it's going to be hugely expensive (this is done for _all_
accesses!) and there's the risk of infinite loops.